+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [SSOButtons] - from presets to Forms/Actions/Button/presets/SSOButtons
 - [Menu] - from Overlays/Menu to Forms/Actions/Menu
 - renaming InformationDisplay to DataDisplay
+- [MapControls] - from presets to Forms/Actions/MapControl/presets/MapControls
 
 ## [v2.131.0] - 09-10-2025